Improve Virtual Memory and ZRAM Swap dynamicly generated in SunlightOS.

+++ kernel/src/process/address_space.rs
@@ -1313,9 +1313,8 @@ impl AddressSpace {
                                 stats.user_frames += 1;
                             }
                         } else if p1e.addr().as_u64() != 0 {
-                            let _ = crate::memory::zram::discard_block(
-                                ((p1e.addr().as_u64() >> 12) - 1) as usize,
-                            );
+                            let _ =
+                                crate::memory::zram::discard_block((p1e.addr().as_u64() >> 12) - 1);
                             stats.swap_blocks += 1;
                         }
                         p1e.set_unused();
